Extract duplicated code to a class
